changing exception location
