+++ model-engine/model_engine_server/api/docker_image_batch_job_bundles_v1.py
@@ -15,12 +15,12 @@ from model_engine_server.common.dtos.batch_jobs import (
 )
 from model_engine_server.common.dtos.model_bundles import ModelBundleOrderBy
 from model_engine_server.core.auth.authentication_repository import User
-from model_engine_server.core.domain_exceptions import (
+from model_engine_server.core.loggers import filename_wo_ext, make_logger
+from model_engine_server.domain.exceptions import (
+    EndpointResourceInvalidRequestException,
     ObjectNotAuthorizedException,
     ObjectNotFoundException,
 )
-from model_engine_server.core.loggers import filename_wo_ext, make_logger
-from model_engine_server.domain.exceptions import EndpointResourceInvalidRequestException
 from model_engine_server.domain.use_cases.docker_image_batch_job_bundle_use_cases import (
     CreateDockerImageBatchJobBundleV1UseCase,
     GetDockerImageBatchJobBundleByIdV1UseCase,
