+++ model-engine/model_engine_server/common/datadog_utils.py
@@ -17,3 +17,9 @@ def add_trace_request_id(request_id: str):
     current_span = tracer.current_span()
     if current_span:
         current_span.set_tag("launch.request_id", request_id)
+
+
+def get_request_id():
+    """Gets the request id for an api call (in our case, dd trace id) so that we can filter in Datadog easier"""
+    current_span = tracer.current_span()
+    return current_span.trace_id if current_span else None
