+++ model-engine/model_engine_server/domain/exceptions.py
@@ -1,4 +1,62 @@
-from model_engine_server.core.domain_exceptions import DomainException
+from dataclasses import dataclass
+
+
+class DomainException(Exception):
+    """
+    Base class for exceptions thrown for domain (business logic) errors.
+    """
+
+
+class ObjectAlreadyExistsException(DomainException):
+    """
+    Thrown when the user tries to create a model with a name that already exists.
+    """
+
+
+class ObjectNotFoundException(DomainException):
+    """
+    Thrown when a required object is not found, e.g. when creating a version for a nonexistent model
+    """
+
+
+class ObjectNotAuthorizedException(DomainException):
+    """
+    Thrown when a user tries to access an object they don't own.
+    """
+
+
+class ObjectHasInvalidValueException(DomainException, ValueError):
+    """
+    Thrown when a user tries to create an object with an invalid value.
+    """
+
+
+class ObjectNotApprovedException(DomainException):
+    """
+    Thrown when a required object is not approved, e.g. for a Bundle in review.
+    """
+
+
+@dataclass
+class DockerImageNotFoundException(DomainException):
+    """
+    Thrown when a user tries to specify a custom Docker image that cannot be found.
+    """
+
+    repository: str
+    tag: str
+
+
+class DockerBuildFailedException(DomainException):
+    """
+    Thrown if the server failed to build a docker image.
+    """
+
+
+class ReadOnlyDatabaseException(DomainException):
+    """
+    Thrown if the server attempted to write to a read-only database.
+    """
 
 
 class ExistingEndpointOperationInProgressException(DomainException):
@@ -112,3 +170,13 @@ class TriggerNameAlreadyExistsException(DomainException):
     """
     Thrown if the requested name already exists in the trigger repository
     """
+
+
+class InternalError(DomainException):
+    """
+    Thrown as a catch-all for unhandled errors.
+    """
+
+    def __init__(self, request_id, error):
+        self.request_id = request_id
+        self.error = error
