+++ model-engine/model_engine_server/domain/use_cases/batch_job_use_cases.py
@@ -17,17 +17,17 @@ from model_engine_server.common.dtos.batch_jobs import (
 )
 from model_engine_server.common.resource_limits import validate_resource_requests
 from model_engine_server.core.auth.authentication_repository import User
-from model_engine_server.core.domain_exceptions import (
-    DockerImageNotFoundException,
-    ObjectHasInvalidValueException,
-    ObjectNotAuthorizedException,
-    ObjectNotFoundException,
-)
 from model_engine_server.core.loggers import filename_wo_ext, make_logger
 from model_engine_server.domain.authorization.live_authorization_module import (
     LiveAuthorizationModule,
 )
 from model_engine_server.domain.entities import ModelEndpointType
+from model_engine_server.domain.exceptions import (
+    DockerImageNotFoundException,
+    ObjectHasInvalidValueException,
+    ObjectNotAuthorizedException,
+    ObjectNotFoundException,
+)
 from model_engine_server.domain.gateways import CronJobGateway, DockerImageBatchJobGateway
 from model_engine_server.domain.repositories import (
     DockerImageBatchJobBundleRepository,
