+++ model-engine/model_engine_server/domain/use_cases/docker_image_batch_job_bundle_use_cases.py
@@ -8,12 +8,12 @@ from model_engine_server.common.dtos.batch_jobs import (
 )
 from model_engine_server.common.dtos.model_bundles import ModelBundleOrderBy
 from model_engine_server.core.auth.authentication_repository import User
-from model_engine_server.core.domain_exceptions import (
-    ObjectNotAuthorizedException,
-    ObjectNotFoundException,
-)
 from model_engine_server.domain.authorization.live_authorization_module import (
     LiveAuthorizationModule,
+)
+from model_engine_server.domain.exceptions import (
+    ObjectNotAuthorizedException,
+    ObjectNotFoundException,
 )
 from model_engine_server.domain.repositories import DockerImageBatchJobBundleRepository
