+++ model-engine/model_engine_server/domain/use_cases/trigger_use_cases.py
@@ -14,16 +14,17 @@ from model_engine_server.common.resource_limits import validate_resource_request
 from model_engine_server.common.settings import REQUIRED_ENDPOINT_LABELS
 from model_engine_server.core.auth.authentication_repository import User
 from model_engine_server.core.config import infra_config
-from model_engine_server.core.domain_exceptions import (
+from model_engine_server.domain.authorization.live_authorization_module import (
+    LiveAuthorizationModule,
+)
+from model_engine_server.domain.exceptions import (
+    CronSyntaxException,
     DockerImageNotFoundException,
+    EndpointLabelsException,
     ObjectHasInvalidValueException,
     ObjectNotAuthorizedException,
     ObjectNotFoundException,
 )
-from model_engine_server.domain.authorization.live_authorization_module import (
-    LiveAuthorizationModule,
-)
-from model_engine_server.domain.exceptions import CronSyntaxException, EndpointLabelsException
 from model_engine_server.domain.gateways.cron_job_gateway import CronJobGateway
 from model_engine_server.domain.repositories import (
     DockerImageBatchJobBundleRepository,
