+++ model-engine/tests/unit/domain/test_docker_image_batch_job_bundle_use_cases.py
@@ -5,7 +5,7 @@ from model_engine_server.common.dtos.batch_jobs import (
 )
 from model_engine_server.common.dtos.model_bundles import ModelBundleOrderBy
 from model_engine_server.core.auth.authentication_repository import User
-from model_engine_server.core.domain_exceptions import (
+from model_engine_server.domain.exceptions import (
     ObjectNotAuthorizedException,
     ObjectNotFoundException,
 )
