+++ model-engine/tests/unit/domain/test_llm_use_cases.py
@@ -14,16 +14,14 @@ from model_engine_server.common.dtos.llms import (
 )
 from model_engine_server.common.dtos.tasks import SyncEndpointPredictV1Response, TaskStatus
 from model_engine_server.core.auth.authentication_repository import User
-from model_engine_server.core.domain_exceptions import (
-    ObjectHasInvalidValueException,
-    ObjectNotAuthorizedException,
-    ObjectNotFoundException,
-)
 from model_engine_server.domain.entities import ModelEndpoint, ModelEndpointType
 from model_engine_server.domain.exceptions import (
     EndpointUnsupportedInferenceTypeException,
     InvalidRequestException,
     LLMFineTuningQuotaReached,
+    ObjectHasInvalidValueException,
+    ObjectNotAuthorizedException,
+    ObjectNotFoundException,
 )
 from model_engine_server.domain.use_cases.llm_fine_tuning_use_cases import (
     MAX_LLM_ENDPOINTS_PER_INTERNAL_USER,
